run simulation

run simulation

Usage

internal_simulation_cpp(
  nDays,
  demographicData,
  temperatureData,
  movementData,
  incubationPeriod,
  infectionDuration,
  initInfectedHumans,
  initInfectedMosquitoes
)

Arguments

nDays

Number of days to simulate.

demographicData

Data frame containing the Demographic data for the synthetic population.

temperatureData

Data frame with the temperature historic.

movementData

Data frame with the movement flow between territories.

incubationPeriod

Incubation period of the disease.

infectionDuration

Infection duration of the disease.

initInfectedHumans

Percentage of initially infected humans.

initInfectedMosquitoes

Percentage of initially infected mosquitoes.

Value

Dataframe with susceptible, infected an recovered individuals each day