Setup ABM data and basic parameters

Setup ABM data and basic parameters

Usage

setup(
  demographic_data,
  temperature_data,
  movement_data = NULL,
  demographic_data_MGN = TRUE,
  divipola_code = NULL,
  disease = "dengue",
  incubation_period = 4,
  infection_duration = 14,
  init_infected_humans = 0.005,
  init_infected_mosquitoes = 0.03
)

Arguments

demographic_data

sf object with the demographic data in the specified format. It's expected to be input by the user from the DANE´s MGN (Marco Geoestadístico Nacional).

temperature_data

Data frame with the temperature historic. It's expected to be input from the historic data at IDEAM´s historic temperature

movement_data

Data frame with the movement flow between territories. If not provided it's automatically calculated from the demographic data using distance and population.

demographic_data_MGN

Boolean stating if the data is directly downloaded from MGN and needs filtering and cleaning.

divipola_code

Divipola code of the municipality to be simulated

disease

Character string for the disease name. Can be either "dengue", "zika" or "chikungunya".

incubation_period

Incubation period of the disease.

infection_duration

Infection duration of the disease.

init_infected_humans

Percentage of initially infected humans.

init_infected_mosquitoes

Percentage of initially infected mosquitoes.

Value

S3 ABM_model object with the saved initial parameters.

Examples

if (FALSE) { # \dontrun{
setup()
} # }